// src/generated/models/Crf82_iatattendancesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crf82_iatattendancescrf82_attendancestatus = {
  0: 'Present',
  1: 'Optional'
} as const;
export type Crf82_iatattendancescrf82_attendancestatus = keyof typeof Crf82_iatattendancescrf82_attendancestatus;
export const Crf82_iatattendancesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crf82_iatattendancesstatecode = keyof typeof Crf82_iatattendancesstatecode;
export const Crf82_iatattendancesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crf82_iatattendancesstatuscode = keyof typeof Crf82_iatattendancesstatuscode;

export interface Crf82_iatattendancesBase {
  crf82_attemptdate?: string;
  crf82_attemptnumber?: number;
  crf82_attempttitle: string;
  crf82_attendancestatus?: Crf82_iatattendancescrf82_attendancestatus;
  crf82_campusname?: string;
  crf82_coursename?: string;
  crf82_gradelevel?: string;
  crf82_iatattendanceid: string;
  crf82_legacyidentifier?: number;
  crf82_slotidentifier?: number;
  crf82_stepcode?: string;
  crf82_studentname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crf82_iatattendancesstatecode;
  statuscode?: Crf82_iatattendancesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Crf82_iatattendances extends Crf82_iatattendancesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  crf82_attendancestatusname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}